Revert "integrate loopback"
