+++ .gitignore
@@ -32,18 +32,3 @@ server/data
 server/auth/auth.js
 
 **/test.*
-*.csv
-*.dat
-*.iml
-*.out
-*.sublime-*
-*.swo
-*.swp
-*.tgz
-*.xml
-.DS_Store
-.idea
-.project
-.strong-pm
-npm-debug.log
-
