--- .yo-rc.json
@@ -1,3 +1,0 @@
-{
-  "generator-loopback": {}
-}
