--- common/models/party.js
@@ -1,3 +1,0 @@
-module.exports = function(Party) {
-
-};
