--- common/models/party.json
@@ -1,28 +1,0 @@
-{
-  "name": "party",
-  "plural": "parties",
-  "base": "PersistedModel",
-  "idInjection": true,
-  "options": {
-    "validateUpsert": true
-  },
-  "properties": {
-    "name": {
-      "type": "string",
-      "required": true
-    },
-    "owner": {
-      "type": "string"
-    },
-    "playlist": {
-      "type": [
-        "any"
-      ],
-      "required": true
-    }
-  },
-  "validations": [],
-  "relations": {},
-  "acls": [],
-  "methods": {}
-}
