--- oldServer/config/database.js
@@ -1,5 +1,0 @@
-module.exports = {
-
-    'url' : process.env.MONGOLAB_URI || 'mongodb://localhost/jibe'
-
-};
