+++ package.json
@@ -22,7 +22,6 @@
     "gulp-karma": "0.0.4",
     "gulp-nodemon": "^1.0.4",
     "gulp-shell": "^0.2.7",
-    "jshint": "^2.5.6",
     "jshint-stylish": "^0.2.0",
     "karma": "^0.12.16",
     "karma-chrome-launcher": "^0.1.4",
@@ -50,12 +49,9 @@
     "bower": "^1.4.1",
     "browser-sync": "^1.5.7",
     "clone-stats": "0.0.1",
-    "compression": "^1.0.3",
     "connect-flash": "~0.1.1",
     "cookie-parser": "~1.0.0",
-    "cors": "^2.5.2",
     "ejs": "~0.8.5",
-    "errorhandler": "^1.1.1",
     "expect.js": "^0.3.1",
     "express": "^4.4.5",
     "express-session": "~1.0.0",
@@ -86,10 +82,6 @@
     "karma-nyan-reporter": "0.0.43",
     "karma-phantomjs-launcher": "^0.1.4",
     "karma-unicorn-reporter": "^0.1.4",
-    "loopback": "^2.14.0",
-    "loopback-boot": "^2.6.5",
-    "loopback-connector-mongodb": "^1.13.0",
-    "loopback-datasource-juggler": "^2.19.0",
     "method-override": "~1.0.0",
     "mocha": "^1.20.1",
     "mongoose": "^4.1",
@@ -99,10 +91,6 @@
     "passport-local": "^1.0.0",
     "passport-spotify": "latest",
     "q": "^1.0.1",
-    "serve-favicon": "^2.0.1",
     "shortid": "^2.2.2"
-  },
-  "optionalDependencies": {
-    "loopback-explorer": "^1.1.0"
   }
 }
