--- server/boot/_root.js
@@ -1,6 +1,0 @@
-module.exports = function(server) {
-  // Install a `/` route that returns server status
-  var router = server.loopback.Router();
-  router.get('/', server.loopback.status());
-  server.use(router);
-};
