--- server/boot/authentication.js
@@ -1,4 +1,0 @@
-module.exports = function enableAuthentication(server) {
-  // enable authentication
-  server.enableAuth();
-};
