--- server/boot/rest-api.js
@@ -1,4 +1,0 @@
-module.exports = function mountRestApi(server) {
-  var restApiRoot = server.get('restApiRoot');
-  server.use(restApiRoot, server.loopback.rest());
-};
