--- server/datasources.local.js
@@ -1,9 +1,0 @@
-var mongoUri = require('./config/database.js').url;
-
-module.exports = {
-  mongodb: {
-    defaultForType: 'mongodb',
-    connector: 'loopback-connector-mongodb',
-    url: mongoUri
-  }
-};
