--- server/middleware.json
@@ -1,30 +1,0 @@
-{
-  "initial:before": {
-    "loopback#favicon": {}
-  },
-  "initial": {
-    "compression": {},
-    "cors": {
-      "params": {
-        "origin": true,
-        "credentials": true,
-        "maxAge": 86400
-      }
-    }
-  },
-  "session": {},
-  "auth": {},
-  "parse": {},
-  "routes": {},
-  "files": {
-    "loopback#static": {
-      "params": "$!../client"
-    }
-  },
-  "final": {
-    "loopback#urlNotFound": {}
-  },
-  "final:after": {
-    "errorhandler": {}
-  }
-}
