+++ views/home.ejs
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html>
+
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title>JIBE | Dominate the Playlist</title>
+        <meta name="description" content="Dominate the playlist">
+        <link rel="stylesheet" href="assets/css/style.css">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+    </head>
+    <body>
+        <div class="navbar">
+            <div class="nav-logo"></div>
+            <a href="/logout"><div class="nav-toggle"><i class="fa fa-sign-out"></i></div></a>
+        </div>
+
+        <div class="logo-container">
+            <div class="logo"></div>
+        </div>
+
+        <div class="button-container">
+
+            <a href="/auth/spotify"><div  class="start-button dark-button button">
+                <span class="bold">START</span><span class="normal"> A PARTY</span>
+            </div></a>
+
+        <div class="start-button light-button button">
+            <span class="bold">JOIN</span><span class="normal"> A PARTY</span>
+        </div>
+
+        </div>
+
+    </body>
+</html>
