+++ views/host.ejs
@@ -1,0 +1,60 @@
+<!doctype html>
+<html ng-app='djBooth'>
+<!-- setting dependencies -->
+<script href='../lib/angular/angular.js'></script>
+<script src="../lib/angular-route/angular-route.js"></script>
+<script src="../lib/jquery/dist/jquery.min.js"></script>
+<script src="../lib/bootstrap/dist/js/bootstrap.min.js"></script>
+<script src="../lib/underscore/underscore.js"></script>
+<!-- UI select components -->
+<script src="../lib/angular-ui-select/dist/select.js"></script>
+<script src="../lib/angular-sanitize/angular-sanitize.js"></script>
+<!-- JS components -->
+<script type="text/javascript" src="../JS/app.js"></script>
+<script type="text/javascript" src="../JS/directives.js"></script>
+<script type="text/javascript" src="../JS/controllers.js"></script>
+<script type="text/javascript" src="../JS/factories.js"></script>
+ <link rel="stylesheet" href="//netdna.bootstrapcdn.com/bootstrap/3.0.2/css/bootstrap.min.css">
+<head>
+    <meta charset="utf-8" />
+    <title>Spotisound</title>
+</head>
+
+<body  class=" jumbotron center-block text-center">
+    <div class='container-fluid' ng-controller='playListController'>
+        <h1>#PartyRoom</h1>
+        <div id="player" ng-controller='playListController'></div>
+        <div ng-controller="YouTubeCtrl">
+          <youtube width="{{yt.width}}" height="{{yt.height}}" videoid="{{yt.videoid}}"></youtube>
+          <button ng-click='newVideo()'>test</button> 
+        </div>
+
+        <!-- search field -->
+        <ui-select ng-model='result.selected' ng-controller='searchController'  on-select='selectSong(result.selected)'  theme="bootstrap" ng-disabled="false" reset-search-input="false" uiSelectConfig.appendToBody = true; style="width: 300px;">
+            <ui-select-match placeholder="Enter a song...">{{$select.selected.title}}</ui-select-match>
+            <ui-select-choices repeat="result in results |limitTo:10" refresh="getSongs($select.search)" refresh-delay="0" >
+                <div ng-bind-html="result.title |highlight: $select.search"></div>
+            </ui-select-choices>
+        </ui-select>
+
+        <!-- playlist -->
+        <div class='panel panel-primary' ng-controller='songController'>
+            <div class='panel-heading'>
+                <h3 class='panel-title'> Playlist</h3>
+            </div>
+            <ul class='panel body list-unstyled'>
+                <!-- 'this is where the playlist is' -->
+                <li ng-repeat="song in playList" class="row-fluid inline ">
+                    <p>
+                        <div>{{song.title}} - {{song.votes}}
+                            <span  ng-click='upVote(song.id)'class="glyphicon glyphicon-thumbs-up" aria-hidden="true"></span>
+                            <span ng-click='downVote()'class="glyphicon glyphicon-thumbs-down" aria-hidden="true"></span>
+                        </div>
+                    </p>
+                </li>
+            </ul>
+        </div>
+    </div>
+</body>
+
+</html>
