+++ views/index.ejs
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta charset="utf-8">
+        <meta http-equiv="X-UA-Compatible" content="IE=edge">
+        <title>JIBE | Dominate the Playlist</title>
+        <meta name="description" content="Dominate the playlist">
+        <meta name="viewport" content="user-scalable=no,width=device-width" />
+        <link rel="stylesheet" href="assets/css/style.css">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+    </head>
+    <body>
+
+        <div class="login-top login-height">
+            <div class="logo"></div>
+            <p class="slogan">Dominate the playlist</p>
+            <a href="/auth/facebook" class="fa-facebook-container">
+                <i class="fa fa-facebook circle fa-2x"></i>
+            </a>
+        </div>
+
+        <div class="login-bottom">
+            <p class="login-label">Login</p>
+        </div>
+
+        <div class="more-info">
+            <p class="more-info-label">Tell me more about Jibe</p>
+        </div>
+
+    </body>
+</html>
