+++ views/profile.ejs
@@ -1,0 +1,79 @@
+<!-- views/profile.ejs -->
+<!doctype html>
+<html>
+<head>
+    <title>Jibe</title>
+    <link rel="stylesheet" href="//netdna.bootstrapcdn.com/bootstrap/3.0.2/css/bootstrap.min.css">
+    <link rel="stylesheet" href="//netdna.bootstrapcdn.com/font-awesome/4.0.3/css/font-awesome.min.css">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.4.0/css/font-awesome.min.css">
+    <style>
+        body        { padding-top:80px; word-wrap:break-word; }
+    </style>
+</head>
+<body>
+<div class="container">
+
+    <div ng-controller='playListController' class="page-header text-center">
+        <h1><span class="fa fa-anchor"></span> Jibe </h1>
+        <a ng-click='newQueue("<%= user._id %>")' class="btn btn-default btn-sm">Create Room</a>
+        <a href="/logout" class="btn btn-default btn-sm"> Logout </a>
+    </div>
+
+    <div class="row">
+
+        <!-- FACEBOOK INFORMATION -->
+        <div class="col-sm-6">
+            <div class="well">
+
+
+                <!-- check if the user has this token (is the user authenticated with this social account) -->
+                <% if (user.facebook.token) { %>
+                    <h3 class="text-primary"><img src='<%= user.facebook.photo %>' class="img-circle" style="height:50px; width: 50px"></span> Facebook</h3>
+                    <p>
+                        <strong>name</strong>: <%= user.facebook.name %><br>
+                        <strong>id</strong>: <%= user.facebook.id %><br>
+                        <strong>token</strong>: <%= user.facebook.token %><br>
+                        <strong>email</strong>: <%= user.facebook.email %><br>
+
+                    </p>
+
+                    <a href="/unlink/facebook" class="btn btn-primary" style="border: none">Unlink</a>
+                <% } else { %>
+                    <h3 class="text-primary"><span class="fa fa-facebook"></span> Facebook</h3>
+                    <a href="/connect/facebook" class="btn btn-primary" style="border: none">Connect Facebook</a>
+                <% } %>
+
+            </div>
+        </div>
+
+
+        <!-- SPOTIFY INFORMATION -->
+        <div class="col-sm-6">
+            <div class="well">
+
+
+                <!-- check if the user has this token (is the user authenticated with this social account) -->
+                <% if (user.spotify.token) { %>
+                    <h3 class="text-info" style="color:#81b71a"><img src='<%= user.spotify.photo %>' class="img-circle" style="height:50px; width: 50px"> Spotify </h3>
+                    <p>
+                        <strong>name</strong>: <%= user.spotify.name %><br>
+                        <strong>id</strong>: <%= user.spotify.id %><br>
+                        <strong>token</strong>: <%= user.spotify.token %><br>
+                        <strong>email</strong>: <%= user.spotify.email %><br>
+
+                    </p>
+
+                    <a href="/unlink/spotify" class="btn btn-info" style="background-color:#81b71a; border: none">Unlink</a>
+                <% } else { %>
+                    <h3 class="text-primary" style="color:#81b71a"><i class="fa fa-spotify"></i> Spotify</h3>
+                    <a href="/connect/spotify" class="btn btn-info" style="background-color:#81b71a; border: none">Connect Spotify</a>
+                <% } %>
+
+            </div>
+        </div>
+
+    </div>
+
+</div>
+</body>
+</html>
